General test cleanup
